+++ CopilotKit/packages/react-core/CHANGELOG.md
@@ -1,5 +1,35 @@
 # ui
 
+## 1.8.9
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.9
+- @copilotkit/shared@1.8.9
+
+## 1.8.9-next.0
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.9-next.0
+- @copilotkit/shared@1.8.9-next.0
+
+## 1.8.8
+
+### Patch Changes
+
+- dfb67c3: - refactor: rename mcpEndpoints to mcpServers for naming consistency
+  - doc changes
+  - @copilotkit/runtime-client-gql@1.8.8
+  - @copilotkit/shared@1.8.8
+
+## 1.8.8-next.1
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.8-next.1
+- @copilotkit/shared@1.8.8-next.1
+
 ## 1.8.8-next.0
 
 ### Patch Changes