+++ CopilotKit/packages/react-ui/CHANGELOG.md
@@ -1,5 +1,38 @@
 # ui
 
+## 1.8.9
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.9
+- @copilotkit/react-core@1.8.9
+- @copilotkit/shared@1.8.9
+
+## 1.8.9-next.0
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.9-next.0
+- @copilotkit/react-core@1.8.9-next.0
+- @copilotkit/shared@1.8.9-next.0
+
+## 1.8.8
+
+### Patch Changes
+
+- Updated dependencies [dfb67c3]
+  - @copilotkit/react-core@1.8.8
+  - @copilotkit/runtime-client-gql@1.8.8
+  - @copilotkit/shared@1.8.8
+
+## 1.8.8-next.1
+
+### Patch Changes
+
+- @copilotkit/runtime-client-gql@1.8.8-next.1
+- @copilotkit/react-core@1.8.8-next.1
+- @copilotkit/shared@1.8.8-next.1
+
 ## 1.8.8-next.0
 
 ### Patch Changes