+++ CopilotKit/packages/runtime-client-gql/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @copilotkit/runtime-client-gql
 
+## 1.8.9
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.9
+
+## 1.8.9-next.0
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.9-next.0
+
+## 1.8.8
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.8
+
+## 1.8.8-next.1
+
+### Patch Changes
+
+- @copilotkit/shared@1.8.8-next.1
+
 ## 1.8.8-next.0
 
 ### Patch Changes